% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffm2pd_function.R
\name{ffm2pd}
\alias{ffm2pd}
\title{Five Factor Model (FFM) to Personality Disorder (PD)}
\usage{
ffm2pd(ffmprofiles)
}
\arguments{
\item{ffmprofiles}{A dataframe with 5 columns (Order: Extraversion-Agreeableness-Concscientiousness-Neuroticism-Openness) and n rows (n=number of individuals). Values represent z Scores on each FFM dimension.}
}
\description{
This function calculates Personality Disorder Risk for one or more FFM personality profiles. The calculation is based on two earlier studies: (1) A meta-analysis by Samuel and Widiger (2008) on the average FFM scores for each DSM Personality Disorder; (2) a recent study by Steppan et al. (2019) on a mathematical link function between the FFM and Personality Disorders. Personality Disorder risk is calculated based on a vector model. Given a multivariate normal distribution of the Five Factor Model, the average distance of individuals to personality disorder prototype vectors can be calculated. Individuals which are significantly closer to a prototype have a higher risk of a given personality disorder.
}
\references{
Samuel, D. B., & Widiger, T. A. (2008). A meta-analytic review of the relationships between the five-factor model and DSM-IV-TR personality disorders: A facet level analysis. Clinical psychology review, 28(8), 1326-1342.
}
\keyword{DSM}
\keyword{Disorder,}
\keyword{FFM,}
\keyword{Personality}
